Filter out discovered superclasses
In general, test cases will be subclasses of the class TestCase (A). But users might
want to subclass a subclass of TestCase instead (B). In that case, neither
class A nor B should be included in the set of discovered test cases.

Before this patch, only class A was being filtered out.